test: split the Squirrel.Mac updater specs and run them serially on x64 (#53663)
* test: split the Squirrel.Mac updater specs and run them serially on x64

Measured on the Intel runners, running two updater tests at once gained
under a minute per suite: each run saturates the three cores, so they took
turns, and every run's timing became less predictable. Run them one at a
time there, and split the suite into two files, the update lifecycle and
the rules for refusing or altering an update, so the shard balancer puts
them on different runners. The harness they share moves to spec/lib.

A slot minted for a retry is dropped when the retry ends instead of
joining the pool, so a retry no longer raises the file's concurrency for
good. The abort-while-running test polls for ShipIt in a loop rather than
on an interval, so a slow `ps` cannot pile up behind itself.

The weights for the two files are estimates from serial runs on the Intel
runners; regenerate them from CI once this has run.

* test: wait for focus, not show, when focusing an inactive panel

On macOS the show event is emitted when the window reports itself
unoccluded, and a panel shown while another app is frontmost may never do
so; the test then timed out waiting for it on the arm64 MAS shard once
api-app-spec ran ahead of it. The panel becoming key is what the test is
about, so wait for focus. Also do not assume the app is active before
handing activation to Finder, and wait for Finder to be frontmost before
showing the panel.

* build: regenerate spec weights with the split updater specs

From Build run 34149004553 on this branch. The two updater files
measured 418s and 379s serially on the Intel runners, against the
estimates of 350 and 385; the rest of the weights are refreshed with them.

Also replaces the temporary instrumentation of the inactive-panel test with the fix.

(cherry picked from commit 57ab2fc74d2c3e87902349a792ab46869e833aec)
